Huge change. Separated server and database logic.
Now is possible to use the same sever connection through a unique SSH tunnel
for different database connections in the same server.

This ways is not need anymore to handle the SSH tunnel for each database client.
Because is already being handled by the server session.

Also was improved error handling and debug logs.