Avoid cast DB types to native JS Date Objects
We have some problems on GUI (https://github.com/sqlectron/sqlectron-gui/issues/117)
because the timezone is applied to the date value. Avoid this parse to native JS Object
will ensure is not applying any timezone. Is the exactly value that is stored in the DB.

It is only implemented to PG and MySQL.
MMSQL does not have support to avoid parse to native JS Object.