Fix SSH tunnel local port
The ssh tunnel was using a port variable that was not defined yet.
I don't know how this was working with electron 0.x versions.
But with electron 1.x that was not working. Probably some upgrade in the v8 has made this wrong behaviour does not work anymore.
Well, this code fix that.