Fix connection problem for old PG versions
Remove 'application_name' attribute because it was causing connection
problems in old PG versions. Also this attribute was not really used by
sqlectron. It just would make easier for a users know the connection
process source in the server. Since it seems is not used for anyone I
will just remove it for now.