Fixes #74
Makes sure that postgres table and column names are surrounded in
double-quotes. This is needed when the table or column name contains
upper case letters.

Also, it adds some additional logic to the autocomplete functionality to
automatically surround postgres table and column names with
double-quotes after the complete event has fired.

In order for the autocomplete to show properly for tables that have
double quotes, it was necessary to add both an unquoted and quoted
variant of the table name to this hintOption object. This does clutter
the dialog a bit - it will show both:

tableNameWithQuotes
"tableNameWithQuotes"

without this, however, you will only get autocomplete when typing the
table name - you will not get column name autocompletion hints off that table
name without the quoted variant present as well.