Prompt for parameter placeholder values
[Feature Demo](https://giphy.com/gifs/QvdLEGkLz5poRrTznY)

If you type in a query like this:

```
SELECT * FROM project WHERE id = :id;
```

You will now get a prompt to provide a value for `:id` before the query
is executed.

This works for queries that use this syntax as well:

```
SELECT * FROM project WHERE id = $1 AND state = $2;
```

The following syntax is NOT supported:

```
SELECT * FROM project WHERE id = ? AND state = ?;
```

This would be possible, but I didnt' want to deal with array indexes at
the time. The current implentation is just an easy object lookup, which
makes all the loops and prompts clean and easy.

I have never used vue before, so if this implementation needs some work,
I am happy to do things differently. Thanks for the great project!