fix for postgresql driver for redshift
Amazon Redshift is based on postgresql 8.0, and pg_catalog.pg_type.typarray was added in 8.3, meaning that the query
would cause an exception to be thrown due to el.typarray not existing when used for Redshift. This change gives us a
fallback in the case of an exception on the first query to use one that will work under pg 8.0. Like the original query,
this is based on the sql that is executed when using \dT command within psql, with the addition of the clause on
n.snpname which was added in later versions of psql for \dT.