rework lib/crypto to be lib/uuid
The original file included encrypt/decrypt functions in addition
to the uuidv4 method. These functions were used in sqlectron
for dealing with its config file. beekeeper-studio has gone
a different path there, and so these methods are not used, and
liable to never to be meaningfully used, not to mention they rely
on deprecated functionality of the crypto library. As such, they
are removed, and the remaining uuidv4 function is moved to a more
appropriately named module (since uuid are not even really crypto
to start with).

Signed-off-by: Matthew Peveler <matt.peveler@gmail.com>