Allow unpin table from "pin" icon
Hi,first of all awesome App! i Just installed is a few minutes ago and I have a little suggestion. I was usgin the "pin" in some tables and I was expecting to unpin a table clicking the "yellow pin", but nothing happens. I think it will be nice to have this feature to allow users to "pin / unpin" faster, rather than moving to the upper tab to unpin them from the X"