Translated README.md into Spanish
I translated the readme doc into Spanish! I'm from Spain so it is peninsular Spanish, not Latin American Spanish (just in case there's differences in expressions, etc). I checked and as far as I'm aware there's no existing version of this doc in Spanish, hope it's useful in some way! If it is, then I'd love to help translate any other important documentation as well.