Added Vim Keybindings to the Editor
This commit adds the option for keymaps to the connection interface and
allows a user to select default, or vim. With vim I added some
additional code in the TabQueryEditor so that Esc is not bound to cancel
a query. That way You do not need to use CTRL+[ to exit insert mode.
We may need to bind cancel query to a different keypress, thoughts?