don't mutate array/object into string, do it for viewing/editing
Instead of mutating the array/object into string for the whole table, we
can convert them into string when showing and editing, so we don't lose
the real data type when calling `tabulator.getData()`, especially when
building an sql.

Getting the real data type is important so knex can generate an sql
correctly. E.g. it can build a postgres array when the argument is a
javascript array (not stringified).