cleanups
some cleanups on the tests. adding a new database shouldn't really
change a lot of things from the tests. most of it were column names that
does not match because firebird always return column names in uppercase.
we can clean this mess by telling knex to lowercase the column names.