add keep most recent command in case of duplicates to vimrc
this is common vim behavior, if you have two commands that are the same,
example would be ```nmap gl ^``` and ```nmap gl $``` then whichever
command is found last in the vimrc file will be the command that is
loaded. I have a test to capture this, and also the logic to hit this.
It searches an array instead of a dict, With having a minimal number of
entries, I think this will be fine to do this way. If a vimrc is longer
than 10-15 lines this could get slow.